Enhance login functionality and improve places rendering
Enhancements were made to the login functionality by integrating a loading element providing user feedback during the process. Changes were also made in the rendering of places, removing the creation of mock data in the service, adding a filter to the Discover page, and implementing an ActionSheetController in the place-detail page to facilitate booking.