Add OfferItemComponent to display offers and refactor code
A new component has been added called OfferItemComponent for displaying offers which significantly restructures the presentation of offers in the app. Also, code in several files such as variables.scss and place-detail.page.ts has been refactored to adhere to coding standards and improve readability. The creation and usage of OfferItemComponent provides a more modular way to manage app components.